move communicateCancel, rename wave_task to bluetooth_task
