+++ genki_wave/threading_runner.py
@@ -8,10 +8,11 @@ from serial.threaded import ReaderThread
 
 from genki_wave.constants import BAUDRATE
 from genki_wave.data.writing import get_start_api_package
-from genki_wave.protocols import ProtocolThread, ProtocolAsyncio, wave_task
+from genki_wave.protocols import ProtocolThread, ProtocolAsyncio, bluetooth_task, CommunicateCancel
+
 
 from genki_wave.utils import get_serial_port, get_or_create_event_loop
-from genki_wave.asyncio_runner import producer_bluetooth, CommunicateCancel, consumer
+from genki_wave.asyncio_runner import producer_bluetooth
 
 
 class ReaderThreadSerial(ReaderThread):
@@ -108,7 +109,7 @@ class WaveListener(threading.Thread):
 
     def run(self):
         self.comm = CommunicateCancel()
-        task = wave_task(self.ble_address, self.comm, self.callbacks)
+        task = bluetooth_task(self.ble_address, self.comm, self.callbacks)
         loop = get_or_create_event_loop()
         loop.run_until_complete(task)
